refactor: schema
